% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{abc_smc_prep}
\alias{abc_smc_prep}
\title{ABC-SMC Slurm Model Preparation}
\usage{
abc_smc_prep(model, prior, nsims, summary_stat_target, prior_test = NULL,
  ncores = 2, dist_weights = NULL, alpha = 0.5, ...)
}
\arguments{
\item{model}{a \code{R} function implementing the model to be simulated. It
must take as arguments a vector of model parameter values and it
must return a vector of summary statistics.}

\item{prior}{a list of prior information. Each element of the list
corresponds to a model parameter. The list element must be a vector
whose first argument determines the type of prior distribution.}

\item{nsims}{the number of simulations below the tolerance threshold is
equal to \code{nsims * alpha}.}

\item{summary_stat_target}{a vector containing the targeted (observed)
summary statistics.}

\item{prior_test}{a string expressing the constraints between model
parameters. This expression will be evaluated as a logical expression,
you can use all the logical operators including \code{"<"}, \code{">"},
\ldots{}. Each parameter should be designated with \code{"X1"}, \code{"X2"},
\ldots{} in the same order as in the prior definition. If not provided,
no constraint will be applied.}

\item{ncores}{Number of cores per node (defines a batch size).}

\item{dist_weights}{a vector containing the weights to apply to the distance
between the computed and the targeted statistics. These weights can
be used to give more importance to a summary statistisc for example.
The weights will be normalized before applying them. If not provided,
no weights will be applied.}

\item{alpha}{Number of simulations to retain in waves 1+.}

\item{...}{Additional arguments can be passed depending on the choosen
method (see below).}
}
\description{
This is the first step in the ABC-SMC Slurm workflow.
}
