% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{abc_smc_wave}
\alias{abc_smc_wave}
\title{Simulates the ABC-SMC Slurm Model}
\usage{
abc_smc_wave(input = "data/", wave, batch, save = TRUE,
  outdir = "data/")
}
\arguments{
\item{input}{A character string containing the directory of the output file
from \code{\link{abc_smc_prep}} which should be saved as an RDS file
with the name \code{abc.wave0.rda} (the default), or the object itself.}

\item{wave}{SMC wave number, where the initial wave = 0. In the standard Slurm
workflow, this would get passed in as an environmental variable
\code{wave} from the master bash script.}

\item{batch}{Batch number for the simulation set, which corresponds to the
array number passed in as an environmental Slurm variable
\code{SLURM_ARRAY_TASK_ID}.}

\item{save}{If \code{TRUE}, writes output to an RDS file with the name
\code{abc.waveX.batchY.rda} in the directory specified by \code{outdir},
where \code{X} is the value of \code{wave} and \code{Y} is the value
of \code{batch}.}

\item{outdir}{Path to save the output RDS file if \code{save=TRUE}.}
}
\description{
This is the second step in the ABC-SMC Slurm workflow.
}
