% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiABC-package.R
\docType{package}
\name{EpiABC-package}
\alias{EpiABC-package}
\alias{EpiABC}
\title{EpiABC: performing efficient approximate Bayesian computation sampling
schemes using R}
\description{
Description
}
\details{
\tabular{ll}{ Package: EpiABC Type: Package Version: 1.5 Date: 2015-06-30
License: GPL-3 LazyLoad: yes }
}
\seealso{
\code{\link{ABC_rejection}}, \code{\link{ABC_sequential}},
\code{\link{ABC_mcmc}}, \code{\link{SABC}}, \code{\link{binary_model}},
\code{\link{binary_model_cluster}}
}
\author{
Franck Jabot, Thierry Faure, Nicolas Dumoulin, Carlo Albert
}
\keyword{abc}
\keyword{package}
