% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{get_posterior}
\alias{get_posterior}
\title{Extracts Posterior Distribution for Parameters and Summary Statistics}
\usage{
get_posterior(wave, input = "data/")
}
\arguments{
\item{wave}{If \code{input} is a character string, the wave file that should
be read from that directory.}

\item{input}{Either a character string with the directory to read the wave
files created with \code{\link{abc_smc_wave}} from, or the direct object
itself.}
}
\description{
Once a simulation wave is complete, this function processes the output and
stores it in a format that is useful for numerical analysis and plotting.
}
