% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{merge_abc}
\alias{merge_abc}
\title{Combines Individual Batch Simulation Files for Processing in Next Wave}
\usage{
merge_abc(wave, indir = "data/", outdir = "data/")
}
\arguments{
\item{wave}{SMC wave number, where the initial wave = 0. In the standard Slurm
workflow, this would get passed in as an environmental variable
\code{wave} from the master bash script.}

\item{indir}{Directory containing the batch files output from
\code{\link{abc_smc_wave}} for merging into a single file.}

\item{outdir}{Path to save the output RDS file if \code{save=TRUE}.}
}
\description{
This is the third step in the ABC-SMC Slurm workflow.
}
