% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{plot.abcsmc}
\alias{plot.abcsmc}
\title{Kernel Density Plot for Posterior Distribution of Parameters and Summary Statistics}
\usage{
\method{plot}{abcsmc}(x, type, stats, mean.line = TRUE,
  stats.positive = FALSE, ...)
}
\arguments{
\item{x}{Output from \code{\link{get_posterior}}.}

\item{type}{Character string of \code{type="stats"} for model statistics
or \code{type="param"} for model parameters.}

\item{stats}{Numeric vector of statistics (or parameters) by position in
\code{x} to plot.}

\item{mean.line}{If \code{TRUE}, plot the mean of the stat/parameter distribution
as a blue line.}

\item{stats.positive}{If \code{TRUE}, constrain the lower bound of the kernel
density at 0 for summary statistics plots.}

\item{...}{Additional arguments based to generic \code{plot}.}
}
\description{
Kernel Density Plot for Posterior Distribution of Parameters and Summary Statistics
}
