% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{sbatch_master_abc}
\alias{sbatch_master_abc}
\title{Create sbatch Bash Shell Script ABC-SMC Workflow}
\usage{
sbatch_master_abc(input, nwaves, master.file = "master.sh",
  runsim.file = "runsim.sh", ckpt = FALSE, append = FALSE,
  mem = "55G", walltime = "1:00:00", user, partition.main = "csde",
  partition.ckpt = "ckpt", account.main = "csde",
  account.ckpt = "csde-ckpt")
}
\arguments{
\item{input}{Output object from \code{\link{abc_smc_prep}}.}

\item{nwaves}{Number of waves of SMC to run, not including wave = 0 that is the
starting wave based on ABC rejection.}

\item{master.file}{Name of the output bash shell script file to write. If
\code{""}, then will print to console.}

\item{runsim.file}{Name of the bash shell script file that runs the R simulation
job.}

\item{ckpt}{If \code{TRUE}, use the checkpoint queue to submit jobs. If
numeric, will specify the first X jobs on the grid as non-backfill.}

\item{append}{If \code{TRUE}, will append lines to a previously created shell
script. New simno will either start with value of \code{simno.start}
or the previous value if missing.}

\item{mem}{Amount of memory needed per node within each Slurm job.}

\item{walltime}{Amount of clock time needed per Slurm job.}

\item{user}{Hyak user name.}

\item{partition.main}{Name of primary HPC partition (passed to -p).}

\item{partition.ckpt}{Name of checkpoint HPC partition (passed to -p).}

\item{account.main}{Name of primary account (passed to -A).}

\item{account.ckpt}{Name of checkpoint account (passed to -A).}
}
\description{
Creates a master-level SLURM::sbatch script given a ABC-SMC
             model prepared and output from \code{abc_smc_prep}.
}
