% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{summary.abcsmc}
\alias{summary.abcsmc}
\title{Numerical Summary of Posterior Distribution of Parameters and Summary Statistics}
\usage{
\method{summary}{abcsmc}(input)
}
\arguments{
\item{input}{Output from \code{\link{get_posterior}}.}
}
\description{
Numerical Summary of Posterior Distribution of Parameters and Summary Statistics
}
