% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiABC-package.R
\docType{package}
\name{EpiABC-package}
\alias{EpiABC-package}
\alias{EpiABC}
\title{EpiABC: Approximate Bayesian Computation for EpiModel on High-Performance Computing Clusters}
\description{
Implementation of Approximate Bayesian Computation with Sequential
Monte Carlo Methods (ABC-SMC) optimized for use with EpiModel R package epidemic
simulations running on High-Performance Computing (HPC) cluster environments such
as MPI and Slurm.
}
\details{
\tabular{ll}{
   Package: \tab EpiABC\cr
   Type: \tab Package\cr
   Version: \tab 1.0.0\cr
   Date: \tab 2019-01-14\cr
   License: \tab GPL-3\cr
   LazyLoad: \tab yes\cr
}
}
\keyword{package}
