% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{summary.abcsmc}
\alias{summary.abcsmc}
\title{Numerical Summary of Posterior Distribution of Parameters and Summary Statistics}
\usage{
\method{summary}{abcsmc}(object, ...)
}
\arguments{
\item{object}{Output from \code{\link{get_posterior}}.}

\item{...}{Additional arguments based to generic \code{summary}.}
}
\description{
Numerical Summary of Posterior Distribution of Parameters and Summary Statistics
}
